Tests: add an image-filtering case found during development
This behavior was noted during development (it lead to a test
failure which has been addressed by removing a set_coord call).
The test here checks that

    set_coord()

does not change the mask filter when the coordinate is not changed.

When the coord value is changed - e.g. physical to logical -
then it is unclear what the filter and mask should be, since we
can't convert the region expression to the new coordinate system.
At present we just test what the code does.