Tests: add PHA handling with grouping/quality
Check some basic handling of columns with grouped and
quality-filtered data.

This lead to issue #1438 which we add a test for as a regression
test, and some other cases where it's not clear what the code
should do so we test the current behavior, even though I believe
there's a strong indication the values should be different.