Tests: apply_filter/grouping length checks (issue #1439)
The issue is whether apply_filter - and apply_grouping for the
DataPHA case - check that the input data is the correct length.

At present it is a hap-hazard response.

If there's been no filter applied (and no quality filter for the
DataPHA) case then the routines do not error out (at least if
the data is not grouped for the DataPHA case).

The remaining calls fail, but there's no "systematic" failure -
some will raise a Sherpa error (DataErr), some an explicit
TypeError, and then others that raise Index/ValueErrors from
un-checked code. These tests are written to test the current
behavior, and so pass. We may need to change these tests - to
catch different errors - if we change the code.