Avoid an un-needed coord change when summing up data
This code is used by calc_*_sum2d routines from sherpa.astro.ui
and anyone who uses the relevant sherpa.astro.utils versions
with the same name.

The removed code seems innocuous - we are re-setting the coord
to the same value it had, when set - but it turns out that
depending on how we make some changes, setting coord here could
remove the original spatial filter. As we haven't changed the
coord setting in the routine, we can avoid all this unpleasentness.
An alternative would be for set_coord to do nothing when it can,
but that's a larger change and this change is still useful even
in that situation, to simplify the code. We also no-longer
bother calling notice or notice2d to clear the temporary mask
as we know the first thing we do in the finally block is to
restore the filter/mask.