Tests: add explicit tests for calc_data_sum/calc_data_sum2d
Ensure we have some explicit checks of various combinations of
the calc_data_sum and calc_data_sum2d cases (noting that the
remated calc_*_* commands tend to rely on the same underlying code
so we do not have to repeat this for all the commands).

Of interest is the handlnig of DataPHA objects with a quality array:
all the objects rely on some level of

 - store the current filter
 - reset the filter
 - apply the new filter
 - sum up the data
 - restore the original filter

and for the DataPHA case this includes ignoring the quality settings,
which is probably not what we want to do here, but working this out
is a different issue and I just wanted to make sure we have tests
of the existing behavior (these tests were created as I noted that
the quality-handling part of the code above had not been triggered
in our existing test suite, which is not surprising as we don't
really have many explicit cases of data like this).