Tests: ensure DataIMG.get_img is fully tested
Coverage tests showed that the case of model + filter was not
being covered, so add tests to cover this. This lead to finding
issue #1447 so I made sure we covered "all" the bases here
(e.g. if the data is not filtered, partially filtered, all
filtered).

Some of these tests should be considered to be regression tests
(those related to issue #1447).