Tests: allow test of str(DataIMGInt) to pass
As noted by @hamogu in review - this test was still marked as
xfail even though it now does work, because the output was not
as I had expected (this is one of the disadvantages to xfail,
as it's hard to say "this fails in this specific way"...)/

I was expecting to get x0lo, x1lo, x0hi, x1hi output but we get
the DataIMG form instead. Rather than mark this as xfail let's
treat this as a regression test so we can find out when if
we ever change it.