data: the independent axis can not change size, once set
The data objects are now restricted so that, once the independent
axis is set, you can not change the size of it, and hence any of
the related fields. That means that code like

    d = Data1D('x', [1, 2, 3], [20, 30, 15])
    d.set_indep(([3, 5, 9], ))

will still work, but

    d.set_indep(([3, 5], ))

will now fail with a DataErr call saying

    independent axis can not chage size (3 to 2).

Users will now have to explicitly copy fields over to a new
data object if they want to re-size the data.

The size attribute has been added to the Data class, and it
returns either None (when no independent axis has been set),
or the number of elements in the independent axis (when set).