Tests: tweak stats data check
Change the test so that it creates a new DataPHA object - with
one-less element than the background - rather than editing the
existing data objcet in place. Both approaches are currently
valid, but creating a new DataPHA object has less issues for
long-term manitenance, particularly if some form of input
validation is added to the data classes.