Tests: ensure sherpa smoke tests are done in a clean environment
This ensures that each smoke test runs in a clean environment,
that is, sherpa.astro.ui.clean() has been called before running
the tests.

Without this you are subject to other tests that may have changed
the statistic (e.g. to Cash) leading to the test failing. This was
only a problem if pytest-xdist was used to run tests on multiple
cores and you happen to hit certani conditions (depending on how
many tests there are, their ordering, and how many cores were used
with pytest-xdist). However, it makes sense to make sure we have
a stable test here.