BUILD with setuptools before version 60
There are various issues when using a recent setuptools version

- changes to distutils
- strange behavior (see issue #1456)

so for now restrict the version in use. This follows the advice
from NumPy:
https://numpy.org/devdocs/reference/distutils_status_migration.html#numpy-setuptools-interaction

There are a number of changes to the CI runs to ensure we
have an old-enough version as it does seem to matter what the
intalled version is (i.e. separate from any pyproject constraint).