GH use develop for the full-conda builds
The current structure of the GitHub actions can lead to the
developer tests (e.g. `python setup.py test` / `pytest`) being
run with an "install" build. This is wasteful, because the
install build is ignored and the code is re-built (thanks to
the use of develop by the test command), and is also going to
cause problems as we simplify the build infrastructure to
reflect PEP 517.

We may well want to have a "full build" with an installed, rather
than developer, build, but let's worry about that in a later PR.