data: ensure data and model dimemsions match (fix issue #1460)
In PR #900 I added the ndim attribute to models, which means that
we can now stop invalid expressions such as

    Const1D + Gauss2D

However, we have not extended this to the data class. This commit
adds the ndim attribute to the Data classes - BUT NOT THE
"simultaneous data" class - which allows us to check that we
are using the "correct" model and data.

There are two principle places where we can check that the two
match

- when creating a sherpa.fit.Fit object
- when evaluating a model to match a data object
  (the eval_model/eval_model_to_fit methods)

but there are several other places which are needed to catch
corner cases (in particular from the UI layer but a user of the
object interface could come across thes). At present the 1D
get_x and get_y routines check this (the 2D routines tend to
use the parent Data routines and so handle this differently).
Unfortunately the use of the get_evaluation_indep routine for the
1D case complicates things (and it's not clear why this is not
also done for the 2D case but that's a different issue). There
may be some corner cases where we are missing this check.

Now, we do not require the ndim attributes, but only check if both
are set. This allows

- for running with a Sherpa session that was created with a 4.14.0
  release or earlier (e.g. saved in a pickle file), where the
  attributes do not exist
- to support user-classes that may not have been updated to match
  the new code (for both models and data classes)
- and there are some cases - e.g. the simulataneous classes - which
  do not have a ndim attribute
- our interface does not make it clear that a model is required,
  just a callable (I think it is actually advantageous for us
  just to follow a simple interface, such as "this is callable",
  rather than say "we must be a model" for most of these routines)

The above discussion is about the object interface, but how does
this interact with the UI interface? Well, it is a bit awkward
because in a sequence like

    ui.load_arrays(...)
    ui.set_source(...)
    ... do something ...

then when do we want to do the size validation? We have applied
validation at the set_source call (actually, set_model as
set_source is an alias for this routine), but it does not catch
all cases. For example, you can

    set a 1D dataset
    set a 1D model
    set a 2D dataset

and currently the last step, which changes the dataset to be 2D,
does not clear the model expression. Perhaps it should, but we
don't have a clear identification of all the fields that are
"linked to" the data object [*], so we currently rely on
downstream checks to catch this case (essentially relying on
the Fit and eval_model checks I described above) to catch
problems such as calc_stat or plot_model after the "state"
has got confused.

[*] that is, we have not made the same identification as we
    have done with the "related fields" for the Data classes