Tests: tweak bin_lo/hi and grouping/filtering handling
Add a couple of tests to check whether bin_lo and bin_hi validate
their lengths (they do not). We have separate routines added
for bin_lo/bin_hi compared to the other fields since they
appear to have different rules (e.g. the number of elements
does not have to match the independent axis), but it's hard to
be sure because there is no meannigful documentation.

Several tests created a DataPHA object with scalar bin_lo/hi
values (added as part of #766) which at that time did nothing
(and does nothing here), but will cause a problem once data
validation is added. So they are removed here (as the settings
are not used in the changed tests).

As part of this update the validation checks (which currently fail)
so that the handling of scalar versus sequence arguments - which
were added to track #1160 - are now done separately, even moving
to different test files, as it highly likely they will error out
differently once data validation is added.

A couple of low-level checks of the sherpa.astro.ui interface
for grouping/quality are added since the logic in these routines
should be moved to the DataPHA class as part of data validation.