Docs: fix up documentation of ui.set_grouping
The documentation suggested that setting the grouping column
automatically set the groupng flag, but it does not. A test has
been added to check this and the documentation updated. A few
other tests have been added to check the set_grouping/quality
handling.

We could change the behavior to match the old documentation.