Tests: check what the array fields of a Data object contain
Check whether we store the supplied values or convert to some
other form (i.e. a ndarray) for the "main" fields of the Data
objects: independent, dependent, statistical and systematic errors,
mask, and for the PHA case grouping, quality, areascal, and
backscal.

These are regression tests, and so can expect to change as the
behavior of the data object is changed.