data: tweak the handling of DataSpaceND
When adding basic support for masked arrays in #793 a check
in the independent axis of the DataSpaceND was made. However,
the argument here (`indep`) is a tuple, not an array-like
object, so the check is meaningless as written. Instead we
should call _check_nomask on every element in the tuple.

There are likely no meaningful tests of this.