Tests: check PHA related-field handling of a set
The idea is to give an argument that triggers potentially-surprising
code paths. In this case what happens if the argument can be converted
to a ndarray but is not array-like, and also "not a number".

The current code does not validate the fields here, so this is a
regresion check.

We could add many more checks like this, but it is expected to be
an edge case.