data: ensure independent axis and related fields match size
Add a _related_fields entry to the Data structure to indicate what
fields should have the same size as the independent axis.

The main complication is in the DataPHA class, because there are
extra fields - such as grouping and backscal - and the channel/counts
aliases. These are now explicit properties that may apply data
type conversions (depending on the property, so grouping and quality
now force OGIP int16 behavior).

The bin_lo and bin_hi fields are completely undocumented, but do not
fit into the concept of a "related" field because they do not have to
match the "size" of the data set (at least, there's at least one test
that has a different size for these compared to the channel array; I
believe the test is technically wrong but I have not investigated).
See issue #1471.

The DataPHA.apply_filter used to call the super-class, but this
no-longer works because of the length validation made there, and
the fact that for the DataPHA case the mask has become disconnected
from the channel array (thanks to grouping). The code could try to
be clever in the super-class, but it does not seem worth it here.
One thought is whether we need an "effective size" attribute which
would allow the code to query both the raw and grouped sizes: the
problem with this is that we'd have to retrofit this into all the
code paths, even for non-PHA data, and then work out when we need
to use which "size" attribute.

One small issue is what error message to display to DataPHA users
who call apply_filter (directly or indirectly) with an array that
neither matches the number of channels or the filtered array length
(this condition is only supported for DataPHA). After a variety of
implementations I went with reporting the filtered length rather than
trying to have something that reports both un-filtered and filtered
lengths, as it didn't seem to be useful.