data: ensure DataPHA.bin_lo/hi match the other fields - fix #1471
The bin_lo and bin_hi fields are treated as any other related
field for the DataPHA object. Minimal documentation is added, but
it is hard to write more because it is unclear what the meaning is,
or how well tested it is (i.e. when a response is also used).