data: reset the mask/filter when the independent axis is changed
If the independent axis is changed then it makes sense to clear the
mask, as there's no guarantee that the old mask is still valid. As
we now can't change the independent axis apart from setting the
indep field as is done here - that is a user can no longer do
something like

    data.indep[0][2] = 200

it makes things a bit clearer (that is, clearing the mask makes
more sense). As discussed below, this also happens when the
coord setting of a `DataIMG` object is changed, as that causes
the independent axes to be replaced.

A warning is raised when the mask is changed (whether a 1D mask
or a region filter) to warn the user.

If the user wants to keep the mask they will need to do something
like

    omask = data.mask
    data.indep = ...
    data.mask = omask

The DataIMG class adds extra complexity here, since

- there's also a region filter (the _region field) to worry about
  which we handle by over-loading the new _clear_filter method so
  that the _region field also gets removed
- switching coordinate systems is an interesting issue as to
  whether we should change the mask; it depends on what we want
  the region filter to mean. We now say that the region filter
  applies to the previous coordinate system, so changing the
  system clears it. There is a special case here so that if the
  coordinate system doesn't actually change we do not clear the
  filter.