instrument: check PSF kernel and data match dimensionality
Ensure the PSFModel kernel is a Data instance, callable, or None,
and take advantage of that knowledge. This also stores the kernel
dimensionality via the Model's ndim attribute. This means that
we automatically get model-expression checks - e.g.

    psf1(m1) + m2

It is possible for a convolved-model to have a ndim set to None,
when the kernel is a callable (but not a model instance). We do
not seem to have any tests of this functionality (although I have
not run the full test-suite), but there is code that supports a
callable-but-not-model case, so keep it (and it is likely that
we have models that can wrap up a callable without knowing the
dimenstionality so we can not assume we have a ndim with the
current code base).

The size, origin, and center attributes are now forced to match the
ndim attribute, if set. When kernel is changed these fields are
cleared if the dimensions do not match (or are None).

The fold method now checks the data and kernel dimensionality and
will error out with a PSFErr if they do not match.

There are some code-style tweaks as well.