instrument: make model of PSFModel read-only
The model attribute is set by the fold method, so ensure it is
only changed by this by making the attribute a read-only property.
Thanks to NoNewAttributesAfterInit it is still a bit odd, as it
appears that the checks on the model attribute are handled in two
different places, but we have checks for both pathways.