instrument: use EvaluationSpace1D for 1D kernels
PSFModel used to use EvaluationSpace2D even for 1D data.
Fix this, we use use an appropriate evaluation space
(so EvaluationSpace1D as we only support 1D and 2D
convolutions).

Most of the other changes here are code style changes, but
I note a possible problem with the handling of lo/hi values
in the _get_kernel_data method: this is only used with DataIMG
data with WCS and we do not have any tests that exercise this
logic, so it's unclear what is meant to happen here.