data: size of dataset now set by x or y
The size of the dataset is now defined by the first field that is
set, not the independent axis. This is unlikely to change things
for most users, but could for strange cases (in which case the
behavior will hpefully be clearer). This behavior is currently
limited to the independent and dependent axes, but it will be
extended to the related axes/fields too.

As part of this, change the validation of the independent axis so
that it happens when creating the object as well as when setting
the indep attribute.