regrid: split Axis into PointAxis and IntegratedAxis
Split Axis so we have

- Axis, using ABCMeta to set up a class that defines an axis
  (not to be used directly)
- PointAxis is for point / non-integrated axes
- IntegratedAxis is for integrated axes

There are some interface changes which could leak through to the
user (PointAxis no longer has lo and hi attributes, for instance),
but it is not expected that users use this functionality (users,
even internal to Sherpa, are expected to use the properties that
have been adjusted to refer to the correct field).

There is more validation done now, with the use of DataErr rather
than ValueError to better match other parts of the Data code.