Tests: corner cases for set_dep/stat/syserror/filter handling in ui
The UI interface adds some extra logic compared to the basic
functionality of the Data classes, so test it out. As we know the
PHA case goes through different code paths we have tests for Data1D
and DataPHA. We do not bother with exercising other classes.

One option would be to move this functionality into the Data
class, but I have not looked into this deeply.