ui: clean up the code
The dep, error (stat/sys), and filter handling has been moved
to separate routines, so that they can easily be re-used from
the astro layer.

There's an argument that the logic in these set_xxx routines
should be moved into the data class, but this functionality
does seem based around the UI layer and it's not clear it
is worth moving it into the already-crowded Data objects.