data: ensure the mask field is the correct size
The mask attribute is a bit tricky to handle compared to the
related fields, as it will be set to True even if the
independent or dependent axes are not set, and because it is
a bool or array field, it feels like we should not use it to
set the size of a dataset (so setting an empty data object
to a known filter state will now trigger an error rather than
silently working). This is down partly to the fact that the
Filter class does not know anything about the independent
axis it is attached to, and I do not think it prudent to
change this interface here.

The DataPHA case adds some complexity, since the mask is defined
to match the grouped data, so the existing length checks from
the super-class are invalid.