Allow the AASCDM dmRegParse call to be used by the region library
If the ASCDM code is available then allow this to be used to parse
region strings, rather than the version in the regino library. This
adds support for FITS region files as the region library can only
handle ASCII files. This is based on a patch that is used to build
CIAO, but extended to allow this option to be selected from the
setup.cfg file.

With the default options, or if

  sherpa_config.region-use-cxc-parser

is explicitly set to False (actually, any string that does not
match "true" when converted to lower case), then the region module
is built as before: the parse_string routine calls either
regReadAsciiRegion or regParse.

If the setting is instead True then the region module is built
with the USE_CXCDM_PARSER macro set, which means that the
parsing uses the dmRegParse routine instead (and the decision for
what to do with a file is converted to a different input string).

When this option is set then the user has to make sure that

a) ascdm.h is accessible from the region-include_dirs setting
b) libascdm.<whatever> is accessible from the region-lib-dirs
   setting
c) ascdm is added to the region-libraries setting

It is expected that a and b will lead to no change, since these
files are expected to be in the same location as cxcregion.h
and libregion.<whatever>, but users can just add the extra
directories to these settings (separated by spaces) if needed.
This means that what most users would need to set is

    region-use-cxc-parser=True
    region-libraries=region ascdm

It is likely that CIAO will not need to change regino-libraries,
but I can not guarantee this.

I am not completely happy with

1) the use of CPP in _region.cc to select the correct code to
   use, but I can't see a way out of this

2) we could automate this setting - e.g. determine when it is
   possible to use dmRegParse - when configuring the code,
   but I do not believe this is worth the significant extra
   complexity that would result