io: ensure get_header_data is an official method of the astro.io backend
In #307 I added the XSPEC read_xstable_model routine, which used both

  sherpa.astro.io.backend.get_table_data
  sherpa.astro.io.backend.get_header_data

but I did not realise that the latter one was not exported by the
backend. Recent changes (e.g. #1185) have improved the astro I/O backend
code, unifying things, but they did not include get_header_data as an
exported symbol. Since we are relying on this in the XSPEC code, we now
mark this symbol as exported (so that any new backends will be expected
to provide it).

The reason this was found was in testing some code when I had no
I/O backend available. It is possible to get into a case when this
missing symbol from dummy_backend.py caused a problem (but it's not
entirely obvious to me in what condition).

Fix #1483