Tests: add tests of expand_grouped_mask
This routine is written in C++, in sherpa/astro/utils/src/_utils.cc
and is tested implicitly by the grouping tests. This adds some
checks of the routine directly.

At the moment the routine expects (but does not require) the mask
and group arguments to

- be the correct size (that is, the mask equals the number of
  group elements that are >= 0)
- the group data "makes sense" (that is, the first element is >= 0).
- the mask argument is not None or a scalar

This means that a number of cases are covered just to see if they
run, with the thought that they will be changed at some point
when the code changes.