Add validation to expand_grouped_mask
Add validation to check that the arguments to expand_grouped_mask
are sensible:

- 1D arrays (so no scalar inputs; higher dimensions are already
  prohibited by the sherpa Array::init method)
- are not empty (this was already done for mask, so extend to
  the group argument)
- the lengths are correct (that is, the number of mask elements
  equals the number of groups, which is the number of values in
  groups which are >= 0)
- the first element of group is >= 0 (previously the code just
  assumed this)

Technically this is not needed, as the existing code does not
trigger any of these failure cases, but I did manage to create
behavior at one point which I believe was caused by the lack
of validation here.

I have slightly changed how the loop works, as we can add a
special-case for the first element, which then avoids a check
on every other loop, but I doubt the change makes much of a
differemce, either in complexity of the code or the run-time
efficiency.

As I've added documentation it has lead to two new defines
for the C++ interface to allow documentation to be included
for functions and modules.