build: use our own version comparison rather than LooseVersion
The LooseVersion class from distutils is deprecated. The suggested
replacement is packaging.version.Version, which would be an
easy change, but it would require the packaging package to be
available when Sherpa is run. As our version-comparison needs are
very limited we can just use a tuple of integers as the version,
as that has the semantics we need (e.g. we do not need the full
PEP 440 support packaging.version.Version provides).