Tests: ensure matplotlib windows are closed after all tests are run
There are times our tests suffer from

https://github.com/The-Compiler/pytest-xvfb/issues/11

That is, the tests run but because smoething still has a window
open then the xvfb server falls over, causing the tests to appear
to fail, when they haven't. Following the discussion in this issue
this commit adds a session-scoped fixture that ensures that any
matplotlib windows are closed before the tests end.