CI move the conda testing into the workflow
The tests that were run by the test.sh script are now included
directly in the workflow for conda, as it makes it easier to see
how the various options interact (for instance, the test-data and
install-type matrix options have an interesting inter-dependence
while we still use the test command from setup.py, as noted in
PR #1458).

The test.sh script is left in with this commit for reference. It
will be removed.