CI split the xspec and ds9 setup scripts
At the moment we treat "having XSPEC" as a synonym for "also
install DS9". Tehcnically they are separate, so split up the
download scripts as it makes it easier to see what each script
does. It would be possible to set up a CI run which only adds
one of XSPEC and DS9 support, but that is not a high priority.