tests: ensure DS9 window is closed after the tests
Ensure that the DS9 imager is closed after the tests. This is
a bit heavy handed as any test run - such as running the tests
in one file that doesn't require DS9 - will lead to a warning
message being created (as we have to try importing the relevant
code), but the pytest machinery seems to hide this.

This follows the approach from #1497 for closing down matplotlib
windows after the tests are run.