Tests: more explicit checks of what is a valid dataset identifier
The current rules are a little unclear, so try to check the
invalid cases. This code is implicitly checked in a number of
tests, but it's nice to add a simple test (and that runs through
all the known invalid values).

One slight complexity is that the astro Session has a larger
set of invaid names than the basic Session. There are also some
open questions about plot types, such as

    compsource vs source_component
    astrodata

which may have had a difference - in that plot("astrodata") may
have produced a different plot from plot("data") or plot_data(),
but it was never clear what this really meant (no documentation)
and some recent changes - e.g. #906 and #984, maybe others - have
changed how we handle the plot types.