ui: set_xlog/... names now match the get_xxx_plot calls
Rather than use short names like "compsource", use the full name,
like "source_component" for the _plot_types dictionary. Since
this dictionary is used by set_xlog/... to select which plot
objects get changed, support is added for the "shortened"
names, by means of the new _plot_types_alias dictionary, which
just maps from the shortened to new names.

If someone uses the old name then a warning message is
created, telling them the correct label to use. For example:

>>> set_xlog("bkgfit")
WARNING: The argument 'bkgfit' is deprecated and 'bkg_fit' should be used instead

This is done as a warning rather than a Python deprecated message
to make sure users see it.

This does not change the set of names that can not be used as
a dataset identifier. This means that

    compsource
    source_component
    bkgfit

are invalid identifier names, but

    bkg_fit

is still a valid name for an identifier.

Note that source_component and model_component were already
supported, so the new labels added here are (only for the astro
module):

    bkg_chisqr
    bkg_delchi
    bkg_fit
    bkg_model
    bkg_ratio
    bkg_resid
    bkg_source

Note that the documentation for set_xlog and other commands still
point to the plot function as the place that describes the labels
that can be used. This is technically correct, but using the
labels described in plot() will lead to a deprecated warning. The
documentation is not changed here as the handling of plot is about
to be changed (to use these new "longer" names), which will then
make the set_xlog/... and plot documentation consistent.