Use direkt html links instead of [1] and a reference section
This replaces the use of a "References" section with direct URL links
in particular in cases where the reference is just a plain old HTML
website and not a scholary publication.

The purpose of this change is to reduce the number of sphinx warnings
caused the the use of non-unique `[1]` for references see #491
for discussion.

Fix some links as suggested in code review

Co-authored-by: wmclaugh <wmclaugh@cfa.harvard.edu>
Co-authored-by: Doug Burke <dburke.gw@gmail.com>