ui: plot object access via a function
Following image and contour handling, the plot access to
the plot objects is now handled via a function. This means
that the plot objects for most, but not all, plots are
now accessed indirectly, rather than by grabbing a named
field from the session object.

The main way that the plot objects change their behavior
based on the "type" of data is to use a different plot
object if the data is a Data1DInt or DataPHA object. These
are the _plot_type values with multiple arguments, which
are always the second value is for Data1DInt and the third
is for DataPHA. In earlier attempts to handle this I had
written fancy checks that used the MRO of the data class
to identify the plot object to use, but it was always
rather complex. The current approach does require some
discipline, but fortunately this can be checked with tests,
and the extra comlpexity of other approaches turns out to
not provide any real advantages. We can always look at
more-complex schemes in the future, as the separation
of object storeage from access to this data is now mediated
by routines, so making future changes easier.

I did experiment with moving the data check outside of
_get_plotobj, but then this means that the astro session
object can not just insert the DataPHA code checks in one
place, but we'd have to overload all the relevant get_xxx_plot
calls, which we do not have to do with this approach.

The handling of get_plot_source_component is a bit strange,
in particular when dealing with template models and
what should be done for PHA data. One of the earlier commits
added tests to check what the code does, and this was used
here to make sure we retain the behavior, but it is nor
clear to me what the code should be doing. I created
issue #1530 to track this (and related issues) as it is not
directly relevant to this work.

The direct-access fields are retained in this commit; they
will be removed in a later commit.