ui: add some typing annotations
Add some support for typing the sherpa.ui.utils and
sherpa.astro.ui.utils modules.

Most of this just adds type annotations, but there are
several code changes

- the addition of an assert statement to satisfy mypy
- the addition of "abstract" methods in the Plot, Histogram,
  Contour, and Image classes so that mypy knows we can call the
  necessary prepare/prepare_image method. I didn't want to
  mark these as abc.abstractmethod since it's not obvious
  how the ABC metaclass would interact with the NoNewXXX class
  and it didn't seem worth experimenting here
- something similar for sherpa.data.Data to add get_filter[_expr]
  since the _get_show_filter routine needs to call the
  get_filter_expr routine and we have labelled get_data() as
  returning a sherpa.data.Data object

To test out I used

    mypy sherpa/ui/utils.py

which can take some time. Tests were basically done by changing an
annotation to an invalid value - such as float or List[float] - and
then looking to see if mypy noticed a problem. If it didn't then I
didn't make the change, as it suggests we need to add other
annotations first. I do note that mypy does complain about the

    copy_reg.constructor(construct_ufunc)

line, but this is an existing issue as it isn't related to the
code I have added or changed.

This started out only changing private routines - those beginning
with _ - but I have added several callable methods. One reason is
that the current code used to create the sherpa.[astro.]ui modules
(e.g. the one that Sherpa users in CIAO will use) does not copy
over the type attributes, so this will not be seen by CIAO users.