extern: remove auto-generated region-lib code
This is fall out from #1401. It appears that the files were
accidentally added because the .gitignore rule to ignore these
files had not been updated to match the change of version number
for the region library: the rule is now "generic" - it matches
region-* - to avoid this issue happening again.