Tests: add basic tests of ui save_xxx (ascii)
Add some tests suggested by coverage checks. We test the ASCII
output of the save_xxx methods of the session classes.

As the output depends on the available backend, we do not
directly check the output. Instead we check we can read back
in the data and get the expected values.

Of course, in doing this we discover that the ui case can write
out ASCII files but the astro.ui case needs a FITS backend,
even when we are not writing out FITS data. This complicates the
tests (to try and make them behave sensibly). The current
approach works, but ends up replicating functionality in
sherpa.utils.testing.

Note that several tests currently fail.