bugfix: support save_model/source/resid/delchi with no id
The string check was being done to early, before we had the chance
to swap id and filename if the latter was not set. Moving the
check into _save_types, which is where it is in the astro version,
fixes this.

The problem fixed here is shown below:

  >>> from sherpa import ui
  >>> ui.load_arrays(1, [1, 2], [2, 3])
  >>> ui.set_source(ui.const1d.mdl)
  >>> ui.save_model("x.dat", clobber=True)
  Traceback (most recent call last):
  ...
  sherpa.utils.err.ArgumentTypeErr: 'filename' must be a string

The problem is that the methods take (id, filename=None, ...)
with the implication being that if only one argument is given,
it is actually the filename argument and the id should be
set to the default value (1 in the example above). Unfortunately
we checked the filename argument before doing the swap of
id and filename values when filename is None, hence the error
message above. This only affects the sherpa.ui.save_xxx calls (and
not all of them, just model, source, resid, and delchi) but not
sherpa.astro.ui, because that had already been fixed to address
this issue (or was never written to cause this problem).