Tests: add regession test of source/model template component plots
There is not enough tests of the template-model code, in particular
what is meant to happen with model/source plots, both for the
plot_model/source and plot_model/source_component_plot options.

These should be considered regression tests, as I do not know what
the intended behavior is. In particular, the template model handling
with no interpolator is unclear.