ui/utils: add pylint annotations to the code base
The pylint tool can be useful, but it's defaults do not map well
to the Sherpa code base, for a number of reasons. The messages
from pylint can be "turned off" for a variet of cases - e.g. just
a single line, a block, the whole file.

Note that we have existing pylint comments in these files, so it's
not a new feature.