ui: consolidate plot handling
In #1177 access to image objects from the Session class was changed
from direct access (e.g. self._xxx) to using a dictionary
(self._image_types). This commit does the same thing for the
plot objects: the objects are accessed from the _plot_types
dictionary rather than having a set of _data/... fields.

For most users this will make no difference. It is only an issue
for code that directly accesses the Session class. However, these
are labelled private fields so they should be aware that things
could change.

With this commit, images, contours, and plots now use the same
scheme. However, there are four differences between the
contour/image and plot handling:

1. we can have multiple objects associated with a plot type - at
   the moment the second element is for Data1DInt data and,
   for the astro version, the second element is for DataPHA data,
   and this makes the get_<label>_plot calls rather cumbersome;

2. the template-model handling is odd and doesn't quite follow
   this (see #1530);

3. the PHA "fit" plot also requires a different plot object to for
   the model than is provided by get_model_plot (for the fit we
   group the data but for the model we do not group it), and the
   access to these objects are currently via direct access to
   the session field (for the bkg fit plot), or just created
   on-the-fly for the "source" plot;

4. and some plot types - e.g. plot_scatter - still require direct
   access to a session field.