test: add corner-case regresion tests for plot()
Add two regression tests for

a) does plot("compsource", mdl) work
b) does plot("compsource") error out

and repeat for the "compmodel" variant. Whilst we are add it,
add tests foe the "longer" forms of "source_component" and
"model_component".

Both tests are considered regresson tests as they are just to
identify when there has been a change. As the aim is to check
whether the labels are correctly processed, there is no actual
check the call has done anything sensible (for case a). For
case b all we do is check we get the expected error message: it
is a "low-level" error that we do not try and catch in plot(),
so the check will note if this ever changes.

We also ad tests for the "bkgxxx" and "bkg_xxx" plot commands.
At present the former pass and the latter fail, so test this
happens. These are also regression tests. Since the behavior is
expected to change over time (e.g. both pass, then probably have
the "bkgxxx" commands start to fail) the tests are kept separate.