ui: remove energy/photon commands from set_xlog/...
This stops the ability of calling

    set_ylog("energy")
    set_ylinear("photon")

oto change the plots created by

    plot_energy_flux() / get_energy_flux_hist()
    plot_photon_flux() / get_photon_flux_hist()

This was never documented (at least in the Sherpa docstrings)
so is unlikely to be missed. The one case I can thinkg of would
be a user who has used either

    set_ylog()
    set_ylog("all")

and is then surprised the plot no-longer changes.

The current work-around for these users is to call the appropriate
get_xxx_flux_hist call and change the preference directly. For
example

    get_energy_flux_hist(recalc=False).histo_prefs["ylog"] = True

This may seem a minor change with no-obvious improvement for the
user, but

- the current behavior is not documented
- future plans to consolidate the plot() and set_xlog() set of
  commands are actually hard to do with the "energy" and "flux"
  labels as they do not match the rest of the system (e.g.
  there is no get_energy_plot() command)

It is better to make the energy and photon plots be handled like
we do the other "special" commands like plot_scatter() and
plot_trace().