ui: remove models from the global table when clean is called
A call like

    set_source(const1d.bob + gauss1d.fred)

will create "global" symbols bob and fred when the default value
of _model_autoassign_func is used. A call to delete_model_component
will delete these symbols, but calling clean() did not. So we
now pass through the model symbol table and remove them when
clean() is called, as long as the default _model_autoassign_func
is set. The idea is that if the user has changed this then we
should not delete the label. This is not a perfect solution but
the intersection of people calling clean() and changing this
field (with set_model_autoassign_func) is likely to be small.

One possibility would be to avoid the check for the autoassign
func but instead only remove those symbols which are derived from
the model class, so that something like

    set_source(const1d.bob)
    bob = 23

will not remove bob.

Thanks to the calls to clean() this effectively fixes #1527
(although there's no guarantee that other tests can't add back
these symbols).